requirements.txt
anyio @ file:///home/conda/feedstock_root/build_artifacts/anyio_1666191106763/work/dist
certifi==2022.12.7
chardet==3.0.4
click @ file:///D:/bld/click_1666798499870/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
fastapi @ file:///home/conda/feedstock_root/build_artifacts/fastapi_1678668504237/work
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
install==1.3.5
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1663332044897/work
nltk==3.6.5
numpy @ file:///D:/bld/numpy_1675642697443/work
pandas @ file:///D:/bld/pandas_1674136611544/work
pydantic @ file:///D:/bld/pydantic_1678316363092/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==1.0.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1673864280276/work
regex @ file:///D:/bld/regex_1667265210597/work
rfc3986==1.5.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1662051266223/work
starlette @ file:///home/conda/feedstock_root/build_artifacts/starlette-recipe_1678556735617/work
textblob==0.15.3
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1677948868469/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
uvicorn @ file:///D:/bld/uvicorn-split_1678464703804/work
